Refactor bounty management: Change bounty_count column type to double precision in tournament_result_players table to support fractional shares. Implement BountyEliminationEventsCache for tracking bounty elimination events, including methods for saving, retrieving, and deleting events. Update InGameUserStateCache and related services to handle new bounty logic, ensuring accurate player state management. Enhance OpenAPI documentation to reflect changes in bounty elimination and undo operations, improving clarity and functionality in tournament management.

+++ db/schema.sql
@@ -46,7 +46,7 @@ create table if not exists tournament_result_players (
     entry_payment_method      text,
     reentry_by_payment_method text,
     total_reentry_count       integer  not null default 0,
-    bounty_count              integer  not null default 0,
+    bounty_count              double precision not null default 0,
     bonuses                   text,
     custom_bonus_chips        text,
     bounty_kills              text,
